Move reporting message from plugutils.* to reporting.*
