+++ misc/plugutils.h
@@ -82,61 +82,6 @@ plug_get_port_transport (struct script_infos *, int);
 struct script_infos *
 plug_create_from_nvti_and_prefs (const nvti_t *);
 
-gchar *
-make_table_driven_lsc_info_json_str (const char *, const char *, const char *,
-                                     const char *, const char *, const char *);
-
-/*
- * Reporting functions
- */
-
-typedef enum
-{
-  ERRMSG,
-  HOST_START,
-  HOST_END,
-  LOG,
-  HOST_DETAIL,
-  ALARM,
-  DEADHOST,
-  HOSTS_COUNT
-} msg_t;
-
-void
-proto_post_alarm (const char *, struct script_infos *, int, const char *,
-                  const char *, const char *);
-
-void
-post_alarm (const char *, struct script_infos *, int, const char *,
-            const char *);
-
-void
-post_alarm_udp (struct script_infos *, int, const char *, const char *);
-
-#define post_alarm_tcp post_alarm
-
-void
-proto_post_error (const char *, struct script_infos *, int, const char *,
-                  const char *, const char *);
-void
-post_error (const char *, struct script_infos *, int, const char *,
-            const char *);
-
-#define post_error_tcp post_error
-
-void
-proto_post_log (const char *, struct script_infos *, int, const char *,
-                const char *, const char *);
-
-void
-post_log (const char *, struct script_infos *, int, const char *);
-
-void
-post_log_with_uri (const char *, struct script_infos *, int, const char *,
-                   const char *);
-
-#define post_log_tcp post_log
-
 /*
  * Management of the portlists
  */
